07/12/2023-Create database & seeding data

+++ webShop.Data/Entities/Cart.cs
@@ -8,8 +8,8 @@ namespace webShopSolution.Data.Entities
 {
     public class Cart
     {
-        public int Id { set; get; }
-        public int ProductId { set; get; }
+        public long Id { set; get; }
+        public long ProductId { set; get; }
         public int Quantity { set; get; }
         public decimal Price { set; get; }
