+++ webShop.Data/Entities/CategoryTranslation.cs
@@ -8,8 +8,8 @@ namespace webShopSolution.Data.Entities
 {
     public class CategoryTranslation
     {
-        public int Id { set; get; }
-        public int CategoryId { set; get; }
+        public long Id { set; get; }
+        public long CategoryId { set; get; }
         public string Name { set; get; }
         public string SeoDescription { set; get; }
         public string SeoTitle { set; get; }
