+++ webShop.Data/Entities/OrderDetail.cs
@@ -8,8 +8,8 @@ namespace webShopSolution.Data.Entities
 {
     public class OrderDetail
     {
-        public int OrderId { set; get; }
-        public int ProductId { set; get; }
+        public long OrderId { set; get; }
+        public long ProductId { set; get; }
         public int Quantity { set; get; }
         public decimal Price { set; get; }
