+++ webShop.Data/Entities/ProductInCategory.cs
@@ -8,11 +8,11 @@ namespace webShopSolution.Data.Entities
 {
     public class ProductInCategory
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         public Product Product { get; set; }
 
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
 
         public Category Category { get; set; }
     }
