+++ webShop.Data/Entities/ProductTranslation.cs
@@ -8,8 +8,8 @@ namespace webShopSolution.Data.Entities
 {
     public class ProductTranslation
     {
-        public int Id { set; get; }
-        public int ProductId { set; get; }
+        public long Id { set; get; }
+        public long ProductId { set; get; }
         public string Name { set; get; }
         public string Description { set; get; }
         public string Details { set; get; }
